refactor(lib): remove net-ts SDK and related scripts from the project and cleanup some unused files in lib crate
- Deleted the `sdks/net-ts` directory, including its package.json, scripts, and configuration files.
- Updated `CLAUDE.md` and `pnpm-workspace.yaml` to reflect the removal of the network-specific TypeScript SDK.
- Adjusted Makefile to remove references to the deleted SDK.
- Cleaned up related documentation and tests to ensure consistency across the codebase.
