+++ crates/lib/src/mod.rs
@@ -1,15 +1,12 @@
 pub mod account;
 pub mod args;
-pub mod cache;
 pub mod config;
 pub mod constant;
 pub mod error;
-pub mod jup;
 pub mod log;
 pub mod oracle;
 pub mod rpc;
 pub mod signer;
-pub mod solana;
 pub mod state;
 pub mod token;
 pub mod transaction;
