--- sdks/net-ts/package.json
@@ -1,29 +1,0 @@
-{
-    "name": "@kora/net",
-    "version": "0.1.0",
-    "description": "TypeScript SDK for Kora Net",
-    "main": "dist/src/index.js",
-    "type": "module",
-    "types": "dist/index.d.ts",
-    "scripts": {
-        "build": "tsc",
-        "test": "jest",
-        "test:watch": "jest --watch",
-        "test:coverage": "jest --coverage",
-        "lint": "eslint src --ext .ts",
-        "format": "prettier --write \"src/**/*.ts\"",
-        "prepare": "pnpm run build"
-    },
-    "dependencies": {
-        "gill": "^0.8.2",
-        "sas-lib": "1.0.7"
-    },
-    "devDependencies": {
-        "@types/jest": "^29.5.12",
-        "@types/mocha": "^10.0.10",
-        "@types/node": "^20.17.27",
-        "@typescript-eslint/eslint-plugin": "^7.1.1",
-        "@typescript-eslint/parser": "^7.1.1",
-        "dotenv": "^16.4.5"
-    }
-}
