--- sdks/net-ts/tsconfig.json
@@ -1,17 +1,0 @@
-{
-    "compilerOptions": {
-      "target": "ES2022",
-      "module": "NodeNext",
-      "moduleResolution": "NodeNext",
-      "declaration": true,
-      "outDir": "./dist",
-      "strict": true,
-      "esModuleInterop": true,
-      "skipLibCheck": true,
-      "forceConsistentCasingInFileNames": true,
-      "resolveJsonModule": true,
-      "types": ["node", "jest"]
-    },
-    "include": ["src", "scripts"],
-    "exclude": ["node_modules", "dist"]
-  }
