+++ sdks/pnpm-workspace.yaml
@@ -1,3 +1,2 @@
 packages:
-  - 'ts'
-# - 'net-ts'
+  - "ts"
